# vertexcodec: Implement scalar compare+popcnt for SSE3/NEON