This redundantly recomputes the number of sentinel elements in the
source byte group using scalar arithmetics exclusively; the general
strategy is to put 0 or 1 into each nibble, and then add them all into
the top nibble using a 64-bit multiply + shift. The sum doesn't overflow
because mode1/2 would never be used when all elements are sentinel,
since mode3 is always picked in these cases as it's shorter.

The reason why this helps is that the computation is shorter
(latency-wise) than the actual decoding that uses SIMD, and this allows
the CPU to overlap the decoding of consecutive byte groups.

This may regress throughput on in-order CPUs but should be generally
beneficial on CPUs where vector and scalar instructions can co-issue
with a reasonably wide window.

Based on suggestion by @dougallj

Performance for SSSE3:

AMD Ryzen 5900X (gcc): 3.85 GB/s => 4.30 GB/s (+11%)
AMD Ryzen 5900X (msvc): 3.66 GB/s => 3.90 GB/s (+6%)
AMD EPYC 7571 (gcc): 1.56 GB/s => 1.66 GB/s (+6%)
Intel Xeon 8259CL (gcc): 1.87 GB/s => 1.87 GB/s (-)
Intel i9-9980HK (gcc): 2.65 GB/s => 2.76 GB/s (+4%)

Performance for NEON:

Apple M2 (clang): 2.99 GB/s => 3.60 GB/s (+20%)
AWS Graviton1 (gcc): 0.89 GB/s => 0.94 GB/s (+5%)
AWS Graviton2 (gcc): 1.30 GB/s => 1.50 GB/s (+15%)
AWS Graviton3 (gcc): 1.89 GB/s => 2.38 GB/s (+25%)